ongoing

+++ src/checker.rs
@@ -1,8 +1,9 @@
 use std::collections::HashMap;
 
+#[derive(Debug)]
 pub struct CheckResult {
     pub name: String,
-    pub values: HashMap<String, f64>,
+    pub values: HashMap<String, f32>,
     pub labels: HashMap<String, String>,
     pub processes: Vec<u16>
 }
